+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+### Added
+* `kubi.zone/parent-zone` labels are now removed in cases where the delegation has lapsed, or the parent zone no longer exists.
+
 ## 0.3.5
 
 ### Added